+++ lib/date_params.rb
@@ -1,5 +1,6 @@
+require 'date_params/controller_additions'
+
 module DateParams
   autoload :VERSION, 'date_params/version'
-  autoload :ControllerAdditions, 'date_params/controller_additions'
   autoload :Parser, 'date_params/parser'
 end