+++ HISTORY.md
@@ -3,6 +3,7 @@
 * Removed support for specifying the namespace as an array in preference of the the `:namespace` option.
 * Added `datetime_parser` controller method for parsing date and time input fields to a DateTimeobject.
 * Added travis support
+* Removed the need to include DateParams::ControllerAdditions into the controller
 
 # 1.0.0 / 2013-04-23